Fast forwarding to most recent genesis

+++ genesis/vis/rasterizer.py
@@ -1,14 +1,9 @@
+import os
 import gc
 
 import numpy as np
 
 import genesis as gs
-
-try:
-    from genesis.ext import pyrender
-    from genesis.ext.pyrender.constants import RenderFlags
-except:
-    pass
 from genesis.repr_base import RBC
 
 
@@ -19,15 +14,17 @@ class Rasterizer(RBC):
         self._camera_nodes = dict()
         self._camera_targets = dict()
         self._offscreen = self._viewer is None
+        self._renderer = None
 
     def build(self):
         if self._context is None:
             return
 
         if self._offscreen:
-            platform = "egl"
-            if gs.platform == "macOS":
-                platform = "pyglet"
+            from genesis.ext import pyrender
+
+            # if environment variable is set, use the platform specified, otherwise some platform-specific default
+            platform = os.environ.get("PYOPENGL_PLATFORM", "egl" if gs.platform == "Linux" else "pyglet")
             self._renderer = pyrender.OffscreenRenderer(
                 pyopengl_platform=platform, seg_node_map=self._context.seg_node_map
             )
@@ -35,6 +32,8 @@ class Rasterizer(RBC):
         self.visualizer = self._context.visualizer
 
     def add_camera(self, camera):
+        from genesis.ext import pyrender
+
         self._camera_nodes[camera.uid] = self._context.add_node(
             pyrender.PerspectiveCamera(
                 yfov=np.deg2rad(camera.fov),
@@ -125,13 +124,14 @@ class Rasterizer(RBC):
         self._context.jit.update_buffer(buffer_updates)
 
     def destroy(self):
-        if self._offscreen:
+        if self._offscreen and self._renderer is not None:
             self._renderer._platform.make_current()
             for target in self._camera_targets:
                 self._camera_targets[target].delete()
             self._renderer.delete()
             del self._renderer
             gc.collect()
+            self._renderer = None
 
     @property
     def viewer(self):
